Complete Phase 19–24 native language and runtime evolution (#36)
* Define Phase 19 runtime polymorphism architecture

* Stage Phase 19 frontend syntax

* Run Phase 19 frontend migration

* Implement Phase 19 frontend syntax

* Restore standard CI after Phase 19 frontend

* Remove Phase 19 frontend migration helper

* Stage Phase 19 frontend fix

* Run Phase 19 frontend fix

* Fix Phase 19 abstract member parsing

* Restore CI after Phase 19 frontend fix

* Remove Phase 19 frontend fix helper

* Fix runtime polymorphism via single-inheritance lowering

* Restore Phase 19 inheritance compiler support

* Export Phase 19 source for local validation

* Fix Phase 19 source export

* Pause CI for Phase 19 atomic patch staging

* Stage Phase 19 virtual dispatch patch part 00

* Stage Phase 19 virtual dispatch patch part 02

* Stage Phase 19 virtual dispatch patch part 03

* Stage Phase 19 virtual dispatch patch part 04

* Stage Phase 19 virtual dispatch patch part 05

* Stage Phase 19 virtual dispatch patch part 06

* Stage Phase 19 virtual dispatch patch part 07

* Add Phase 19 atomic patch applicator

* Run Phase 19 virtual dispatch migration

* Retry complete Phase 19 virtual dispatch patch

* Stage Phase 19 virtual dispatch patch part 01

* Decode Phase 19 patch fragments independently

* Diagnose Phase 19 patch fragments

* Add Phase 19 fragment repair helper

* Repair and apply Phase 19 virtual patch

* Implement Phase 19 virtual dispatch

* Document Phase 19 virtual dispatch completion

* Manually align Phase19D runtime interface dispatch validation

* Implement RealScript feature updates

* Implement RealScript application updates

* Close cross-platform warnings and bytecode baselines

* Stabilize inheritance ordering and virtual ABI

* Close global object-model contracts

* Document Phase 19 review validation

* Record successful Phase 19 branch validation

---------

+++ tests/phase18_native_control_flow_tests.cpp
@@ -343,7 +343,8 @@ double main()
     uint f = 6;
     ulong g = 7;
     float h = 8.5;
-    return a + b + c + d + e + f + g + h;
+    return (double)a + (double)b + (double)c + (double)d +
+        (double)e + (double)f + (double)g + (double)h;
 }
 )");
     require(
